requirements.txt
requests>=2.31.0
Pillow>=10.0.0
feedparser>=6.0.10
python-dotenv>=1.0.0
